Task 101

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,9 +7,9 @@
  */
 void print_number(int n)
 {
-	long m;
+	unsigned int  m;
 	int c;
-	long num = n;
+	unsigned int  num = n;
 
 	if (num < 0)
 	{
